remove integration test base class
